requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
astroid==3.3.8
attrs==24.3.0
audioop-lts==0.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
dill==0.3.9
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
injector==0.22.0
isort==5.13.2
mccabe==0.7.0
multidict==6.1.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
pylint==3.3.3
pytest==8.3.4
pytest-asyncio==0.25.1
python-dotenv==1.0.1
requests==2.32.3
structlog==24.4.0
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
